+++ src/embed/embed.spec.ts
@@ -20,6 +20,7 @@ beforeAll(() => {
         thoughtSpotHost,
         authType: AuthType.None,
     });
+    spyOn(window, 'alert');
 });
 
 describe('test view config', () => {
@@ -48,6 +49,7 @@ describe('test view config', () => {
     });
 
     test('trying to register event handler after render should throw error', async () => {
+        spyOn(console, 'error');
         const onErrorSpy = jest.fn();
         const searchEmbed = new SearchEmbed(getRootEl(), defaultViewConfig);
         searchEmbed