+++ src/embed/liveboard.spec.ts
@@ -1,6 +1,12 @@
 import { LiveboardEmbed, LiveboardViewConfig } from './liveboard';
 import { init } from '../index';
-import { Action, AuthType, EmbedEvent, RuntimeFilterOp } from '../types';
+import {
+    Action,
+    AuthType,
+    EmbedEvent,
+    HostEvent,
+    RuntimeFilterOp,
+} from '../types';
 import {
     executeAfterWait,
     getDocumentBody,
@@ -8,6 +14,7 @@ import {
     getRootEl,
 } from '../test/test-utils';
 import { version } from '../../package.json';
+import * as processTriggerInstance from '../utils/processTrigger';
 
 const defaultViewConfig = {
     frameParams: {
@@ -16,10 +23,14 @@ const defaultViewConfig = {
     },
 };
 const liveboardId = 'eca215d4-0d2c-4a55-90e3-d81ef6848ae0';
+const activeTabId = '502693ba-9818-4e71-8ecd-d1a194e46861';
 const vizId = '6e73f724-660e-11eb-ae93-0242ac130002';
 const thoughtSpotHost = 'tshost';
-const defaultParams = `&hostAppUrl=local-host&viewPortHeight=768&viewPortWidth=1024&sdkVersion=${version}`;
+const defaultParamsSansHideAction = `&hostAppUrl=local-host&viewPortHeight=768&viewPortWidth=1024&sdkVersion=${version}`;
+const defaultParams = `${defaultParamsSansHideAction}&hideAction=[%22${Action.ReportError}%22]`;
 const prefixParams = '&isLiveboardEmbed=true&isPinboardV2Enabled=false';
+const prefixParamsVizEmbed =
+    '&isLiveboardEmbed=true&isVizEmbed=true&isPinboardV2Enabled=false';
 
 beforeAll(() => {
     init({
@@ -60,7 +71,7 @@ describe('Liveboard/viz embed tests', () => {
         liveboardEmbed.render();
         await executeAfterWait(() => {
             expect(getIFrameSrc()).toBe(
-                `http://${thoughtSpotHost}/?embedApp=true${defaultParams}&disableAction=[%22${Action.DownloadAsCsv}%22,%22${Action.DownloadAsPdf}%22,%22${Action.DownloadAsXlsx}%22]&disableHint=Action%20denied${prefixParams}#/embed/viz/${liveboardId}`,
+                `http://${thoughtSpotHost}/?embedApp=true${defaultParamsSansHideAction}&disableAction=[%22${Action.DownloadAsCsv}%22,%22${Action.DownloadAsPdf}%22,%22${Action.DownloadAsXlsx}%22]&disableHint=Action%20denied&hideAction=[%22${Action.ReportError}%22]${prefixParams}#/embed/viz/${liveboardId}`,
             );
         });
     });
@@ -78,7 +89,7 @@ describe('Liveboard/viz embed tests', () => {
         liveboardEmbed.render();
         await executeAfterWait(() => {
             expect(getIFrameSrc()).toBe(
-                `http://${thoughtSpotHost}/?embedApp=true${defaultParams}&hideAction=[%22${Action.DownloadAsCsv}%22,%22${Action.DownloadAsPdf}%22,%22${Action.DownloadAsXlsx}%22]${prefixParams}#/embed/viz/${liveboardId}`,
+                `http://${thoughtSpotHost}/?embedApp=true${defaultParamsSansHideAction}&hideAction=[%22${Action.ReportError}%22,%22${Action.DownloadAsCsv}%22,%22${Action.DownloadAsPdf}%22,%22${Action.DownloadAsXlsx}%22]${prefixParams}#/embed/viz/${liveboardId}`,
             );
         });
     });
@@ -152,7 +163,7 @@ describe('Liveboard/viz embed tests', () => {
         liveboardEmbed.render();
         await executeAfterWait(() => {
             expect(getIFrameSrc()).toBe(
-                `http://${thoughtSpotHost}/?embedApp=true${defaultParams}${prefixParams}#/embed/viz/${liveboardId}/${vizId}`,
+                `http://${thoughtSpotHost}/?embedApp=true${defaultParams}${prefixParamsVizEmbed}#/embed/viz/${liveboardId}/${vizId}`,
             );
         });
     });
@@ -173,7 +184,7 @@ describe('Liveboard/viz embed tests', () => {
         liveboardEmbed.render();
         await executeAfterWait(() => {
             expect(getIFrameSrc()).toBe(
-                `http://${thoughtSpotHost}/?embedApp=true&col1=sales&op1=EQ&val1=1000${defaultParams}${prefixParams}#/embed/viz/${liveboardId}/${vizId}`,
+                `http://${thoughtSpotHost}/?embedApp=true&col1=sales&op1=EQ&val1=1000${defaultParams}${prefixParamsVizEmbed}#/embed/viz/${liveboardId}/${vizId}`,
             );
         });
     });
@@ -206,6 +217,35 @@ describe('Liveboard/viz embed tests', () => {
         await executeAfterWait(() => {
             expect(getIFrameSrc()).toBe(
                 `http://${thoughtSpotHost}/?embedApp=true${defaultParams}&pinboardVisibleVizs=[%22abcd%22,%22pqrs%22]${prefixParams}#/embed/viz/${liveboardId}`,
+            );
+        });
+    });
+    test('should process the trigger, for vizEmbed', async () => {
+        const mockProcessTrigger = spyOn(
+            processTriggerInstance,
+            'processTrigger',
+        );
+        const liveboardEmbed = new LiveboardEmbed(getRootEl(), {
+            enableVizTransformations: true,
+            ...defaultViewConfig,
+            vizId: '1234',
+            liveboardId,
+        } as LiveboardViewConfig);
+        liveboardEmbed.render();
+        const result = await liveboardEmbed.trigger(HostEvent.Pin);
+        expect(mockProcessTrigger).toBeCalled();
+    });
+
+    test('should render active tab when activeTab present', async () => {
+        const liveboardEmbed = new LiveboardEmbed(getRootEl(), {
+            liveboardId,
+            activeTabId,
+            liveboardV2: true,
+        } as LiveboardViewConfig);
+        liveboardEmbed.render();
+        await executeAfterWait(() => {
+            expect(getIFrameSrc()).toBe(
+                `http://${thoughtSpotHost}/?embedApp=true${defaultParams}&isLiveboardEmbed=true&isPinboardV2Enabled=true#/embed/viz/${liveboardId}/tab/${activeTabId}`,
             );
         });
     });