+++ src/embed/liveboard.ts
@@ -35,7 +35,7 @@ export interface LiveboardViewConfig extends ViewConfig {
      * This is the minimum height(in pixels) for a full height Liveboard.
      * Setting this height helps resolves issues with empty Liveboards and
      * other screens navigable from a Liveboard.
-     * @version 1.5.0 or later
+     * @version SDK: 1.5.0 | ThoughtSpot: ts7.oct.cl, 7.2.1
      * @default 500
      */
     defaultHeight?: number;
@@ -49,7 +49,7 @@ export interface LiveboardViewConfig extends ViewConfig {
      */
     liveboardId?: string;
     /**
-     * To support backward compatibilty
+     * To support backward compatibility
      * @hidden
      */
     pinboardId?: string;
@@ -66,7 +66,7 @@ export interface LiveboardViewConfig extends ViewConfig {
      * Array of viz ids which should be visible when the liveboard
      * first renders. This can be changed by triggering the "SetVisibleVizs"
      * event.
-     * @version 1.9.1 or later
+     * @version SDK: 1.9.1 | ThoughtSpot: 8.1.0.cl, 8.4.1-sw
      */
     visibleVizs?: string[];
     /**
@@ -75,13 +75,16 @@ export interface LiveboardViewConfig extends ViewConfig {
      */
     preventPinboardFilterRemoval?: boolean;
     /**
-     * Render embedded liveboards and visualizations using the new v2 rendering mode
-     * This is a transient flag which is primarily meant for internal use
+     * Render embedded Liveboards and visualizations in the new Liveboard experience mode
      * @default false
-     * @version SDK: 1.11.0 | ThoughtSpot: 8.3.0.cl
-     * @hidden
+     * @version SDK: 1.14.0 | ThoughtSpot: 8.6.0.cl, 8.8.1-sw
      */
     liveboardV2?: boolean;
+    /**
+     * Tab Id of the Liveboard that is supposed to be active
+     * @version SDK: 1.15.0 | ThoughtSpot: 8.7.0.cl, 8.8.1-sw
+     */
+    activeTabId?: string;
 }
 
 /**
@@ -110,6 +113,8 @@ export class LiveboardEmbed extends V1Embed {
             defaultHeight,
             visibleVizs,
             liveboardV2 = false,
+            vizId,
+            activeTabId,
         } = this.viewConfig;
 
         const preventLiveboardFilterRemoval =
@@ -134,6 +139,9 @@ export class LiveboardEmbed extends V1Embed {
             params[Param.visibleVizs] = visibleVizs;
         }
         params[Param.livedBoardEmbed] = true;
+        if (vizId) {
+            params[Param.vizEmbed] = true;
+        }
         params[Param.LiveboardV2Enabled] = liveboardV2;
         const queryParams = getQueryParamString(params, true);
 
@@ -152,6 +160,7 @@ export class LiveboardEmbed extends V1Embed {
         liveboardId: string,
         vizId?: string,
         runtimeFilters?: RuntimeFilter[],
+        activeTabId?: string,
     ) {
         const filterQuery = getFilterQuery(runtimeFilters || []);
         const queryParams = this.getEmbedParams();
@@ -164,6 +173,9 @@ export class LiveboardEmbed extends V1Embed {
             false,
             false,
         )}/viz/${liveboardId}`;
+        if (activeTabId) {
+            url = `${url}/tab/${activeTabId}`;
+        }
         if (vizId) {
             url = `${url}/${vizId}`;
         }
@@ -195,12 +207,25 @@ export class LiveboardEmbed extends V1Embed {
     };
 
     /**
+     * Triggers an event to the embedded app
+     * @param messageType The event type
+     * @param data The payload to send with the message
+     */
+    public trigger(messageType: HostEvent, data: any = {}): Promise<any> {
+        const dataWithVizId = data;
+        if (this.viewConfig.vizId) {
+            dataWithVizId.vizId = this.viewConfig.vizId;
+        }
+        return super.trigger(messageType, dataWithVizId);
+    }
+
+    /**
      * Render an embedded ThoughtSpot Liveboard or visualization
      * @param renderOptions An object specifying the Liveboard ID,
      * visualization ID and the runtime filters.
      */
     public render(): LiveboardEmbed {
-        const { vizId, runtimeFilters } = this.viewConfig;
+        const { vizId, activeTabId, runtimeFilters } = this.viewConfig;
         const liveboardId =
             this.viewConfig.liveboardId ?? this.viewConfig.pinboardId;
 
@@ -219,7 +244,12 @@ export class LiveboardEmbed extends V1Embed {
 
         super.render();
 
-        const src = this.getIFrameSrc(liveboardId, vizId, runtimeFilters);
+        const src = this.getIFrameSrc(
+            liveboardId,
+            vizId,
+            runtimeFilters,
+            activeTabId,
+        );
         this.renderV1Embed(src);
 
         return this;