+++ src/utils.spec.ts
@@ -8,6 +8,8 @@ import {
     getCssDimension,
     getEncodedQueryParamsString,
     appendToUrlHash,
+    getRedirectUrl,
+    checkReleaseVersionInBeta,
 } from './utils';
 import { RuntimeFilterOp } from './types';
 
@@ -80,8 +82,61 @@ describe('unit test for utils', () => {
         );
     });
 
+    describe('getRedirectURL', () => {
+        test('Should return correct value when path is undefined', () => {
+            expect(getRedirectUrl('http://myhost:3000', 'hashFrag')).toBe(
+                'http://myhost:3000#hashFrag',
+            );
+            expect(getRedirectUrl('http://xyz.com/#foo', 'bar')).toBe(
+                'http://xyz.com/#foobar',
+            );
+        });
+
+        test('Should return correct value when path is set', () => {
+            Object.defineProperty(window.location, 'origin', {
+                get: () => 'http://myhost:3000',
+            });
+
+            expect(
+                getRedirectUrl('http://myhost:3000/', 'hashFrag', '/bar'),
+            ).toBe('http://myhost:3000/bar#hashFrag');
+
+            expect(
+                getRedirectUrl('http://myhost:3000/#/foo', 'hashFrag', '#/bar'),
+            ).toBe('http://myhost:3000/#/barhashFrag');
+        });
+    });
+
     test('getEncodedQueryParamsString', () => {
         expect(getEncodedQueryParamsString('')).toBe('');
         expect(getEncodedQueryParamsString('test')).toBe('dGVzdA');
+    });
+
+    test('when ReleaseVersion is empty ', () => {
+        expect(checkReleaseVersionInBeta('', false)).toBe(false);
+    });
+
+    test('when ReleaseVersion is 7.0.1.cl ', () => {
+        expect(checkReleaseVersionInBeta('7.0.1.cl', false)).toBe(false);
+    });
+
+    test('when cluster has dev version', () => {
+        expect(checkReleaseVersionInBeta('dev', false)).toBe(false);
+    });
+
+    test('when cluster is above 8.4.0.cl-11 software version', () => {
+        expect(checkReleaseVersionInBeta('8.4.0.cl-117', false)).toBe(false);
+    });
+
+    test('when cluster is bellow 8.0.0.sw software version', () => {
+        expect(checkReleaseVersionInBeta('7.2.1.sw', false)).toBe(true);
+    });
+
+    test('when suppressBetaWarning is true and ReleaseVersion is 7.0.1', () => {
+        expect(checkReleaseVersionInBeta('7.0.1', true)).toBe(false);
+    });
+
+    test('when suppressBetaWarning is false ReleaseVersion is 7.0.1', () => {
+        expect(checkReleaseVersionInBeta('7.0.1', false)).toBe(true);
     });
 });